feat: Create pfunctions and Update .zshrc and aliases
pfunctions adds a layer of compartimentalizations to the .zshrc
management and adds a file for function creation and management that
doesn't dirty the aliases or .zshrc files.

For now, the only function there is cdl, one that does cd for the
argument received and then lists the files below it. Super useful and I
had incorrectly created it as an alias, which I conveniently ignored
until a few minutes ago, which, inconviniently, also didn't do what I asked of
it. A failure on all sides, but apparently a profitable one.