feat: updated init.lua and pre-installs
I am in the process of adding lazyvim to my setup.

In order to do so, a lot of changes have to be done to my current setup
and it being a custom one, it will take proportionately longer given I
am working and in college.

For now, I added a comment to the init.lua where my lazy installation
would go. I still have to check if that's possible.

As for the pre-installs.sh, I have added neovim and a special
configuration so that the unstable ppa is the one to be installed, as I
don't get to use lazyvim without it.

In any case, wish me luck.