feat: Create Lazy Configuration and Update init.lua
It's been a while since I wanted to add lazy. It's been recommended to
me for a while and I've found what I've read good enough. I finally did
it and in a way that doesn't clash with my personal dotfiles
configuration.

As for init.lua, it was getting cramped, so I switched to requiring
files in which I made the configuration to make it all easier to read,
as I tend to do with bash files.

I'm quite proud of the unfinished product that I've built. It is quite
nice, even if incomplete.