feat(api): Update Lazy, Nvim, Kitty and Zsh Configs, as well as .gitignore and connections.sh.
With the acclaimed arrival of LazyVim, I've managed to update my nvim
setup quite a bit. I've been following some written, as well as video,
guides and the documentation of some plugins and I've managed to
upgrade, more than update, my configuration.

For one, the init.lua file is more modular than ever before. It requires
only three files: config.options, config.lazy and
config.commands.shebang. The purpose of this is to granulate the
workings of the system so that each change can be more easily done and
so that each file opened is smaller.

I am quite proud, even if the product isn't quite as excellent as I'd
want it it at the moment.

Kitty and ZSH's configurations have changed for the same reason, I
needed a nerd font for Lazy's proper configuration and for the full
power of most nvim plugins. Looking at you, lualine, don't think I
didn't notice.

As for the connections.sh, I had to modify it in order to make sure that
the new fonts were all available in any machine I went to. Fun, to be
more clear.

Last but not least, the following plugins were added:

- Everforest theme, in the shape of the colorscheme.lua file.
- Lsp-config, in order to properly configure LSPs.
- Lualine, because quite frankly it looks cool and gives a good amount
  of information that is quite useful.
- Mason-lspconfig, to connect Mason and Lspconfig.
- Mason, to easily manage LSP servers. Apparently it can also deal with
  DAP servers, linters and formatters, although I have still to learn
  what those are. Perhaps next week.
- Minipairs, to create a pair of commonly user pairs throughout
  programming, such as quote marks and parenthesis. Quite useful, if I
  must say after using it for so short.
- Neotree, for nice folder and file explorer. I can't say I use it much,
  as I've gotten used to using the terminal, but it is nice to look at
  and doesn't take much processing power.
- Telescope in order to fuzzy find files and words. Quite nice, although
  it's current inhability to replace has me bothered, but I wager I'll
  be able to do something about it, I just don't quite know how yet.
- Todocomments, to search for all current TODO comments. Quite simple
  but I've found it's quite the useful functionality.
- Treesitter, for better highlighting and integration with other tools.
  Technically it is a set of programming parsers that convert the text I
  write into a tree of programming language constructs that are easier
  to interact with. I, however, could not properly explain that if put
  in front of a Jury, so, for now, better highlighting will have to
  suffice.
- Trouble, to search and summarize all current issues that the LSP
  servers might've found within my code. Haven't had much chance to use
  that as the code I'm writing is quite guided, so there's little room
  for error.