feat(api): Update plugins
We continue to have an issue with Neotree not working properly only in
git_status. Started a discussion on the github.

On other changes, modified mason to include luacheck, pre-installs to
install luarocks to be able to install luachek and added autopairs,
which seems I wrote wrong. Will fix that in a second.

It's been a good day, off to work.