feat(api): Create LazyVim Config
Replaced previous personal configuration with LazyVim powered
configuration to ease the need for maintenance on my side.

More configuration is to be done in order to reach similar coverage as
the previous configuration, but so far the change has been good.

Will still not change this to main, but it could be considered a success
already.