feat(config): Update Zsh config, Aliases, Lazy config and pre-installs
Updated .zshrc to include zsh-autocomplete for personal needs. It's been
helpful already. I also moved the sourcing of my files and the plugins
to a different file named sources, sourced it and put it after compinit.
Added fast-syntax-highlighting to my zsh so now everything looks
prettier (in Everforest style)

Added new plugins as git submodules.

Added new aliases to relieve the typing needed to reach certain
directories I use quite often. Feel free to browse, they've already
proved useful. I should do a alias use statistic, it'd be quite
revealing. It'd put my git skills to the question, quite frankly.

Updated Lazy config to include new options and keymaps, specifically
pertaining to spellang and a keymap to not type "z=" all day long, but I
might erase it altogether as it doesn't provide much utility.

Modified pre-installs to make it more efficient and complete. Turned
critical apps into an associative array so that on one side we have the
package name and in the other the executable, to then make it even more
efficient to check if the critical package is installed and even easier
to install it. Altogether quite the big upgrade, if I do say so myself.