feat(api): Updated LazyVim and Zsh
Update LazyVim configuration to include some extras. I'll list them
below:
- AI: Added Copilot and Copilot-Chat
- DAP: Added Core and Lua support. Might add JUnit and Google Test
  support in the future, but it's been a long enough day.
- LSP:
  - Clang for C and C++ (I wanna learn).
  - Docker
  - Git
  - Java
  - JSON
  - Kotlin
  - Markdown
  - Rust
  - SQL
  - YAML

I can attest that at the very least it all works for Java, Docker, YAML and Lua.
Haven't gotten around to playing with most of the rest.

Updated Snacks to modify the dashboard to my personal one. Kind of lame,
can't lie, but it makes me happy, so whatevs.

Updated sources file that is sourced in .zshrc in order to make most of
these LSPs work, so I had to add a new plugin called ZSH-NVM to install
nvm in my system and make it work with my pre-installs file.

Finally got spellang to work on vim, which means I have tons of new
words saved. I'll probably be updating that file almost everyday, but I
feel like not pushing it to my dotfiles would be a waste of my own time.

As part of this entire setup, I added the plugin as part of my git
submodules. Getting used and liking it so much.