feat(feature): Update Shebang command, Aliases and Gitignore
Updated the Shebang command as it had some errors that caused lack of
proper effect.

Updated aliases to include certain aliases that I don't want to make
public.

Updated gitignore to include the *.secrets pattern to allow said file to
be ignored by git.

This will be the last commit of this feature as I intend to merge into
main after I create a copy of main. The change has proved beneficial and
more efficient at accomplishing what we intended to do.