refactor(config): Update dotfiles installation
Due to errors encountered in the server branch installation process on
another machine, I noticed some errors in our install script and sourced
scripts.

A more modular approach was taken, where the process is duly separated
into more parts:
- essential-installs: The essential packages are ensured to be installed
  before the rest of the installation proceeds to assure proper
  functioning of the rest of the process.
- connections: the connections script is now more efficient with it's
  code, space and echos.
- optional-installs: the not so essential but still desired packages are
  installed after all previous installations are made so that no errors
  happen.

In addition to all this, instead of using an associative array which
cannot ensure the order of installation, a JSON config file has been
added with all information necessary to proceed to the installation,
with the respective packages divided in separate arrays to facilitate
installation and looping through the array.

On top of that, to not clout the essential and optional install scripts,
the code for each installation, no matter how simple, is separated into
a separate file with it's own name, making it less cluttered in the
install script, easier to notice and manage errors and, frankly, much
more cool.

Should probably add a README, but who has time.

Modified some file names to make them clearer and more in line with good
practices.