feat(api): Update node manager, change to fnm
As is widely known, nvm is quite the slow program and it causes the
console startup time to be exponentially longer.

After some investigation, it seemed to me the most appropriate solution
was to change from nvm to fnm, a Rust based manager that is considerably
faster.

It took me some time to do the change, but it works correctly now and it
is as fast as needed.