refactor(install): modify install script as well as helper scripts
The refactoring continues. Apparently, I didn't do too good of a job the
first time, so I have quite some refactoring to do. The install script
will be getting more polluted as we include more functions and variables
to include both opts and configuration from the command line. For now,
it is all incomplete, alas I have to work.

Eliminate "sudo apt update" from some scripts to avoid repeating the
update, the logic will be pulled to the install script for better
coordination between installations, as well as an installing from pm
function to avoid code duplication, even if it couples the scripts more.
Will have to modify install.json to reflect this, but didn't have time
today.