feat(api): upgrade argument checking
Added checks to the setup process that ensure that the script has
received arguments, that the script has received a -d|--devenv argument
and that if it did those arguments have values in them (whether the
argument is proper or not will be handled later).

Added some testing to be able to quickly and efficiently test the
installation script. Since this is clearly a temporary solution, I chose
to gitignore and will eventually delete. I just don't quite have the
time to delve into shunit2.

Added some scripts to the dotfiles-installation file, as well.
Preparation for the future. They're still incomplete but they check if
there's an environment file at all, if within it you can find the
environment that the user has just requested and then proceeds to read
the packages in said environment and invoke the install_package function
that I'll have to fill in another day.